move MMEEventsManager+Spy into include folder of CTestHelper
