+++ Sources/CTestHelper/CTestHelper.h
@@ -6,4 +6,4 @@ FOUNDATION_EXPORT double CTestHelperVersionNumber;
 //! Project version string for CTestHelper.
 FOUNDATION_EXPORT const unsigned char CTestHelperVersionString[];
 
-#import <CTestHelper/MMEEventsManager+Spy.h>
+#import "MMEEventsManager+Spy.h"
